fix bug with drag and drop

+++ src/index.js
@@ -1,17 +1,16 @@
-import React, {useState, useRef} from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import Draggable from 'react-draggable';
 import './index.css'
 import {
     ControlledMenu,
-    Menu,
     MenuItem,
-    MenuButton,
-    SubMenu
 } from '@szhsin/react-menu';
 import '@szhsin/react-menu/dist/index.css';
-
 import {SketchField, Tools} from 'react-sketch';
+
+
+const TEXTS_ID= 'texts';
 
 function getTextWidth(text) {
   const canvas = document.createElement('canvas');
@@ -55,7 +54,7 @@ class TextArea extends React.Component{
     this.isEditting = false;
   }
 
-  handleChange(e){
+  handleOnChange(e){
     this.isEditting = true;
     const text = e.target.value;
     const lines = text.split("\n");
@@ -88,11 +87,10 @@ class TextArea extends React.Component{
     this.element.focus();
   }
 
-  handleStopDrag(){
-    const rect = this.element.getBoundingClientRect();
+  handleOnStopDrag(e){
     this.setState({
-      x:rect.x,
-      y:rect.y,
+      x:e.offsetLeft,
+      y:e.offsetTop,
     })
   }
 
@@ -109,14 +107,14 @@ class TextArea extends React.Component{
 
     return (
       <Draggable
-        onStop={this.handleStopDrag.bind(this)}
+        onStop={this.handleOnStopDrag.bind(this)}
       >
         <textarea style={style} autoFocus 
           className={`textarea-purge focus:border-blue-300`}
           ref={(element) => {this.element = element }}
           rows={this.state.rows}
           onBlur={this.handleOnBlur.bind(this)}
-          onChange={this.handleChange.bind(this)}/>  
+          onChange={this.handleOnChange.bind(this)}/>  
 
       </Draggable>
     )
@@ -157,13 +155,15 @@ class Paper extends React.Component {
   }
 
   handleOnClick(e) {
+    if (e.target.id != TEXTS_ID) return;
+    
     if (this.state.intact){
       this.setState({intact:false})
     }
+
     const rect = e.target.getBoundingClientRect();
     const x = e.clientX - rect.left,
       y = e.clientY - rect.top - 10; // a lil - 10 doesn't kill nobody
-    console.log(e.target);
     const divs = this.state.divs.slice();
     const overlappedDivs = divs.filter(div => isInBoundingBox(x, y, 
       div.ref.current.state.x, div.ref.current.state.y, 
@@ -253,7 +253,7 @@ class Paper extends React.Component {
           className="absolute w-screen h-screen overflow-scroll top-0 left-0"
         >
           <div 
-            id="texts"
+            id={TEXTS_ID}
             className={`bg-transparent ${this.getZIndex('text')} w-x2 h-x2 absolute`}
             onClick={this.handleOnClick.bind(this)}
           >
